Revert "Customize multi select"
