+++ src/contexts/customize.tsx
@@ -1,119 +1,17 @@
-import React, {
-  createContext,
-  ReactChild,
-  useState,
-  useMemo,
-  useEffect,
-} from "react"
+import React, { createContext, ReactChild, useState, useMemo } from "react"
 import {
   SelectedVariants,
   ShopifyVariant,
   SavedCustomizeContexts,
-  SelectedVariantStorage,
 } from "../types/global"
 
-const defaultSelectedVariants = {
-  step1: {
-    image: {
-      originalSrc: "",
-      altText: "",
-      localFile: {
-        childImageSharp: {
-          gatsbyImageData: {},
-        },
-      },
-    },
-    legacyResourceId: "",
-    price: "0.00",
-    product: {
-      title: "",
-      description: "",
-      onlineStoreUrl: "",
-      productType: "",
-      collections: {
-        handle: "",
-        title: "",
-      },
-      vendor: "",
-    },
-    selectedOptions: [
-      {
-        name: "",
-        value: "",
-      },
-    ],
-    storefrontId: "",
-    sku: "",
-    title: "",
-  },
-  step2: {
-    image: {
-      originalSrc: "",
-      altText: "",
-      localFile: {
-        childImageSharp: {
-          gatsbyImageData: {},
-        },
-      },
-    },
-    legacyResourceId: "",
-    price: "0.00",
-    product: {
-      title: "",
-      description: "",
-      onlineStoreUrl: "",
-      productType: "",
-      collections: {
-        handle: "",
-        title: "",
-      },
-      vendor: "",
-    },
-    selectedOptions: [
-      {
-        name: "",
-        value: "",
-      },
-    ],
-    storefrontId: "",
-    sku: "",
-    title: "",
-  },
-  step3: {
-    image: {
-      originalSrc: "",
-      altText: "",
-      localFile: {
-        childImageSharp: {
-          gatsbyImageData: {},
-        },
-      },
-    },
-    legacyResourceId: "",
-    price: "0.00",
-    product: {
-      title: "",
-      description: "",
-      onlineStoreUrl: "",
-      productType: "",
-      collections: {
-        handle: "",
-        title: "",
-      },
-      vendor: "",
-    },
-    selectedOptions: [
-      {
-        name: "",
-        value: "",
-      },
-    ],
-    storefrontId: "",
-    sku: "",
-    title: "",
-  },
-  step4: [
-    {
+const defaultContext = {
+  currentStep: 1,
+  setCurrentStep: (currentStep: number) => {},
+  productUrl: "/",
+  setProductUrl: (productUrl: string) => {},
+  selectedVariants: {
+    step1: {
       image: {
         originalSrc: "",
         altText: "",
@@ -124,7 +22,7 @@ const defaultSelectedVariants = {
         },
       },
       legacyResourceId: "",
-      price: "0.00",
+      price: 0,
       product: {
         title: "",
         description: "",
@@ -146,48 +44,139 @@ const defaultSelectedVariants = {
       sku: "",
       title: "",
     },
-  ],
-  case: {
-    image: {
-      originalSrc: "",
-      altText: "",
-      localFile: {
-        childImageSharp: {
-          gatsbyImageData: {},
+    step2: {
+      image: {
+        originalSrc: "",
+        altText: "",
+        localFile: {
+          childImageSharp: {
+            gatsbyImageData: {},
+          },
         },
       },
-    },
-    legacyResourceId: "",
-    price: "0.00",
-    product: {
-      title: "",
-      description: "",
-      onlineStoreUrl: "",
-      productType: "",
-      collections: {
-        handle: "",
+      legacyResourceId: "",
+      price: 0,
+      product: {
         title: "",
+        description: "",
+        onlineStoreUrl: "",
+        productType: "",
+        collections: {
+          handle: "",
+          title: "",
+        },
+        vendor: "",
       },
-      vendor: "",
+      selectedOptions: [
+        {
+          name: "",
+          value: "",
+        },
+      ],
+      storefrontId: "",
+      sku: "",
+      title: "",
     },
-    selectedOptions: [
-      {
-        name: "",
-        value: "",
+    step3: {
+      image: {
+        originalSrc: "",
+        altText: "",
+        localFile: {
+          childImageSharp: {
+            gatsbyImageData: {},
+          },
+        },
       },
-    ],
-    storefrontId: "",
-    sku: "",
-    title: "",
+      legacyResourceId: "",
+      price: 0,
+      product: {
+        title: "",
+        description: "",
+        onlineStoreUrl: "",
+        productType: "",
+        collections: {
+          handle: "",
+          title: "",
+        },
+        vendor: "",
+      },
+      selectedOptions: [
+        {
+          name: "",
+          value: "",
+        },
+      ],
+      storefrontId: "",
+      sku: "",
+      title: "",
+    },
+    step4: {
+      image: {
+        originalSrc: "",
+        altText: "",
+        localFile: {
+          childImageSharp: {
+            gatsbyImageData: {},
+          },
+        },
+      },
+      legacyResourceId: "",
+      price: 0,
+      product: {
+        title: "",
+        description: "",
+        onlineStoreUrl: "",
+        productType: "",
+        collections: {
+          handle: "",
+          title: "",
+        },
+        vendor: "",
+      },
+      selectedOptions: [
+        {
+          name: "",
+          value: "",
+        },
+      ],
+      storefrontId: "",
+      sku: "",
+      title: "",
+    },
+    case: {
+      image: {
+        originalSrc: "",
+        altText: "",
+        localFile: {
+          childImageSharp: {
+            gatsbyImageData: {},
+          },
+        },
+      },
+      legacyResourceId: "",
+      price: 0,
+      product: {
+        title: "",
+        description: "",
+        onlineStoreUrl: "",
+        productType: "",
+        collections: {
+          handle: "",
+          title: "",
+        },
+        vendor: "",
+      },
+      selectedOptions: [
+        {
+          name: "",
+          value: "",
+        },
+      ],
+      storefrontId: "",
+      sku: "",
+      title: "",
+    },
   },
-}
-
-const defaultContext = {
-  currentStep: 1,
-  setCurrentStep: (currentStep: number) => {},
-  productUrl: "/",
-  setProductUrl: (productUrl: string) => {},
-  selectedVariants: defaultSelectedVariants,
   setSelectedVariants: (selectedVariants: SelectedVariants) => {},
   setSelectedVariantsToDefault: () => {},
   hasSavedCustomized: {
@@ -212,45 +201,298 @@ export const CustomizeProvider = ({ children }: { children: ReactChild }) => {
     step4: false,
     case: false,
   })
-
-  const [selectedVariants, setSelectedVariants] = useState<SelectedVariants>(
-    defaultSelectedVariants
-  )
+  const [selectedVariants, setSelectedVariants] = useState<SelectedVariants>({
+    step1: {
+      image: {
+        originalSrc: "",
+        altText: "",
+        localFile: {
+          childImageSharp: {
+            gatsbyImageData: {},
+          },
+        },
+      },
+      legacyResourceId: "",
+      price: 0,
+      product: {
+        title: "",
+        description: "",
+        onlineStoreUrl: "",
+        productType: "",
+        collections: {
+          handle: "",
+          title: "",
+        },
+        vendor: "",
+      },
+      selectedOptions: [
+        {
+          name: "",
+          value: "",
+        },
+      ],
+      storefrontId: "",
+      sku: "",
+      title: "",
+    },
+    step2: {
+      image: {
+        originalSrc: "",
+        altText: "",
+        localFile: {
+          childImageSharp: {
+            gatsbyImageData: {},
+          },
+        },
+      },
+      legacyResourceId: "",
+      price: 0,
+      product: {
+        title: "",
+        description: "",
+        onlineStoreUrl: "",
+        productType: "",
+        collections: {
+          handle: "",
+          title: "",
+        },
+        vendor: "",
+      },
+      selectedOptions: [
+        {
+          name: "",
+          value: "",
+        },
+      ],
+      storefrontId: "",
+      sku: "",
+      title: "",
+    },
+    step3: {
+      image: {
+        originalSrc: "",
+        altText: "",
+        localFile: {
+          childImageSharp: {
+            gatsbyImageData: {},
+          },
+        },
+      },
+      legacyResourceId: "",
+      price: 0,
+      product: {
+        title: "",
+        description: "",
+        onlineStoreUrl: "",
+        productType: "",
+        collections: {
+          handle: "",
+          title: "",
+        },
+        vendor: "",
+      },
+      selectedOptions: [
+        {
+          name: "",
+          value: "",
+        },
+      ],
+      storefrontId: "",
+      sku: "",
+      title: "",
+    },
+    step4: {
+      image: {
+        originalSrc: "",
+        altText: "",
+        localFile: {
+          childImageSharp: {
+            gatsbyImageData: {},
+          },
+        },
+      },
+      legacyResourceId: "",
+      price: 0,
+      product: {
+        title: "",
+        description: "",
+        onlineStoreUrl: "",
+        productType: "",
+        collections: {
+          handle: "",
+          title: "",
+        },
+        vendor: "",
+      },
+      selectedOptions: [
+        {
+          name: "",
+          value: "",
+        },
+      ],
+      storefrontId: "",
+      sku: "",
+      title: "",
+    },
+    case: {
+      image: {
+        originalSrc: "",
+        altText: "",
+        localFile: {
+          childImageSharp: {
+            gatsbyImageData: {},
+          },
+        },
+      },
+      legacyResourceId: "",
+      price: 0,
+      product: {
+        title: "",
+        description: "",
+        onlineStoreUrl: "",
+        productType: "",
+        collections: {
+          handle: "",
+          title: "",
+        },
+        vendor: "",
+      },
+      selectedOptions: [
+        {
+          name: "",
+          value: "",
+        },
+      ],
+      storefrontId: "",
+      sku: "",
+      title: "",
+    },
+  })
 
   const setSelectedVariantsToDefault = () => {
-    setSelectedVariants(defaultSelectedVariants)
+    setSelectedVariants({
+      step1: {
+        image: {
+          altText: "",
+          localFile: {
+            childImageSharp: {
+              gatsbyImageData: {},
+            },
+          },
+        },
+        price: 0,
+        product: {
+          title: "",
+          description: "",
+        },
+        selectedOptions: [
+          {
+            name: "",
+            value: "",
+          },
+        ],
+        storefrontId: "",
+        sku: "",
+        title: "",
+      },
+      step2: {
+        image: {
+          altText: "",
+          localFile: {
+            childImageSharp: {
+              gatsbyImageData: {},
+            },
+          },
+        },
+        price: 0,
+        product: {
+          title: "",
+          description: "",
+        },
+        selectedOptions: [
+          {
+            name: "",
+            value: "",
+          },
+        ],
+        storefrontId: "",
+        sku: "",
+        title: "",
+      },
+      step3: {
+        image: {
+          altText: "",
+          localFile: {
+            childImageSharp: {
+              gatsbyImageData: {},
+            },
+          },
+        },
+        price: 0,
+        product: {
+          title: "",
+          description: "",
+        },
+        selectedOptions: [
+          {
+            name: "",
+            value: "",
+          },
+        ],
+        storefrontId: "",
+        sku: "",
+        title: "",
+      },
+      step4: {
+        image: {
+          altText: "",
+          localFile: {
+            childImageSharp: {
+              gatsbyImageData: {},
+            },
+          },
+        },
+        price: 0,
+        product: {
+          title: "",
+          description: "",
+        },
+        selectedOptions: [
+          {
+            name: "",
+            value: "",
+          },
+        ],
+        storefrontId: "",
+        sku: "",
+        title: "",
+      },
+      case: {
+        image: {
+          altText: "",
+          localFile: {
+            childImageSharp: {
+              gatsbyImageData: {},
+            },
+          },
+        },
+        price: 0,
+        product: {
+          title: "",
+          description: "",
+        },
+        selectedOptions: [
+          {
+            name: "",
+            value: "",
+          },
+        ],
+        storefrontId: "",
+        sku: "",
+        title: "",
+      },
+    })
   }
-
-  useEffect(() => {
-    const isBrowser: boolean = typeof window !== "undefined"
-    if (isBrowser) {
-      const urlParams = new URLSearchParams(window.location.search)
-      const custom_id = urlParams.get("custom_id")
-      const customsResume = localStorage.getItem("customs-resume")
-      if (customsResume && custom_id) {
-        const customsStorage = JSON.parse(
-          customsResume
-        ) as SelectedVariantStorage
-        const parsedCustoms = customsStorage.value.customs
-        const resumedSelectedVariants =
-          parsedCustoms[Number(custom_id)].selectedVariants
-        const handle = parsedCustoms[Number(custom_id)].handle
-        const sku = parsedCustoms[Number(custom_id)].sku
-        // prepare context for editing
-        // setting context
-        setSelectedVariants(resumedSelectedVariants)
-        // setting savedCustomized context so radio won't default to top option
-        setHasSavedCustomized({
-          step1: true,
-          step2: true,
-          step3: true,
-          step4: true,
-          case: true,
-        })
-        setCurrentStep(5)
-      }
-    }
-  }, [])
 
   const value = useMemo(
     () => ({
