+++ src/types/customize.d.ts
@@ -147,16 +147,3 @@ export interface ShopifyProductVariant {
     originalSrc: string
   }
 }
-
-export interface rxDetails {
-  sph: string
-  cyl: string
-  axis: string
-  add: string
-  pd: string
-}
-
-export interface rxType {
-  right: rxDetails
-  left: rxDetails
-}
